Bump and don't use Any::Moose
Bump various things here, the perl version, CPAN modules, and remove
the Any::Moose dependency.

Surely I don't need the latest version of perl / CPAN modules. But
this is guaranteed to work, and as see in a TODO here for a test
there's some module changes upstream we still need to work out.